+++ apps/web/src/app/oss.module.ts
@@ -4,10 +4,10 @@ import { LooseComponentsModule } from "./modules/loose-components.module";
 import { OrganizationCreateModule } from "./modules/organizations/create/organization-create.module";
 import { OrganizationManageModule } from "./modules/organizations/manage/organization-manage.module";
 import { OrganizationUserModule } from "./modules/organizations/users/organization-user.module";
-import { PipesModule } from "./modules/pipes/pipes.module";
 import { TrialInitiationModule } from "./modules/trial-initiation/trial-initiation.module";
 import { VaultFilterModule } from "./modules/vault-filter/vault-filter.module";
 import { OrganizationBadgeModule } from "./modules/vault/modules/organization-badge/organization-badge.module";
+import { PipesModule } from "./shared/pipes/pipes.module";
 import { SharedModule } from "./shared/shared.module";
 
 @NgModule({
