+++ apps/web/src/app/shared/pipes/get-organization-name.pipe.ts
